Update readme, rename bvgsensor.py to sensor.py
